Fix test in US time zones
`new Date('2024-01-01')` evaluates to 2024-01-01T00:00:00.000Z, then `getFullYear` resolves that in the local time zone. In time zones that are negative of UTC (such as America/New_York), this results in `getFullYear` giving the previous year, causing a spurious test failure.

Moving the test dates away from January 1 and December 31 fixes this.